@@ -1,11 +1,12 @@
 {
   "name": "beat-timer",
-  "version": "2.3.1",
+  "version": "2.3.2",
   "author": {
     "name": "Samuel Braun",
     "url": "https://www.webry.com/about/"
   },
   "private": true,
+  "type": "module",
   "scripts": {
     "build": "run-p type-check \"build-only {@}\" --",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
@@ -19,17 +20,17 @@
   "dependencies": {
     "@ffmpeg/ffmpeg": "^0.11.6",
     "@vercel/analytics": "^1.1.0",
-    "vue": "^3.3.4",
+    "vue": "^3.5.13",
     "web-audio-beat-detector": "^8.1.54"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.3.3",
     "@tsconfig/node18": "^18.2.2",
     "@types/node": "^18.17.17",
-    "@vitejs/plugin-vue": "^4.3.4",
+    "@vitejs/plugin-vue": "^5.2.1",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/eslint-config-typescript": "^12.0.0",
-    "@vue/tsconfig": "^0.4.0",
+    "@vue/tsconfig": "^0.7.0",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.49.0",
     "eslint-plugin-vue": "^9.17.0",
@@ -40,7 +41,7 @@
     "prettier-plugin-tailwindcss": "^0.5.6",
     "tailwindcss": "^3.3.3",
     "typescript": "~5.2.0",
-    "vite": "^4.4.9",
+    "vite": "^6.0.11",
     "vue-tsc": "^1.8.11"
   }
 }
